Commented out the call to lt_dlclose, since this is apparently the source of the segmentation fault problem.  Actually, it is not lt_dlclose itself that causes the error, but the fact that there are memory leaks elsewhere in the code.  When lt_dlclose is called, those memory management problems make surface.
Notice that in Geoffrey's original dyndrivers code dlclose was never called,
so the bug never manifested itself.  (In other words: the bug was already in
the previous released version of PLplot, and have not been introduced by my
recent changes.)

Of course, this is an interim solution.  This should make PLplot work
properly again, though.

svn path=/trunk/; revision=4322