This is a first try in integrating the DocBook source configuration into the main configuration scheme.  I did not yet try to actually build the documentation, but at least bootstrap.sh and configure work.
There is now a configure option --enable-docbook, which is set to "no" by
default.  If it is set to "yes", then the checks in doc/docbook/configure
are realized and the files under doc/docbook are AC_SUBSTituted.

The final goal of this effort will be to integrate the building of the docs
into the main build and set up things such that all the generated files get
included into the tarball with "make dist" (when that will happen,
make_tarball.sh will be rm -f and we will make a big celebration :-).

Alan, if you are hearing to this, please test when you can.  I am almost
sure it won't work perfectly, but I will try to be responsive on the bug
reports.

svn path=/trunk/; revision=4346