Removed $(BASE).tar.gz from the list of dependencies of $(INDEX_HTML).  This was triggering a call to "make dist", which does not make sense now that the doc/docbook/configure.ac has been integrated into the topdir configure.ac. "make dist" can only be issued at the topdir.
Also, removed $(BASE).tar.gz as one of the elements of $(TARBALL).  This
means that the DocBook source tarball will not be downloaded to the web
site.  Accordingly, the corresponding line has been removed from
www/index.html.in.in.

We should probably mention in the web page that the DocBook sources are
available in the PLplot distribution tarball.

svn path=/trunk/; revision=4349