Since the advent of the Autotools configuration scheme, the replacement of the string PLPLOT_OCTAVE_DIR in file plplot_octave_path.m is not done anymore.  The necessary code to do it was in cf/pkg_octave.in, which was included in one of the old Octave bindings Makefile.
The simplest way that I found to circumvent this problem is to take
advantage of the AC_SUBST mechanism of Autoconf.  Therefore, I cvs removed
the plplot_octave_path.m file and added plplot_octave_path.m.in, which
contains the string @OCTAVE_DIR@ that gets replaced when AC_OUTPUT is
called.

Joao, if you are hearing to this and disagree with my changes, please
manifest yourself.  Without this change, or something functionally
equivalent, the Octave bindings are unusable.

svn path=/trunk/; revision=4369