This is a first experiment to get the different README (and othe documentation ) files that are scattered through the source directory tree into ${datadir}/doc/plplot.
I started by doing the changes for the Octave bindings (please, Joao, check
if this suits you).  The Octave bindings have the following documentation
files: BUGS, FGA, INSTALL, ToDo, USAGE, and README.  They are now installed
in the docdir with the suffix ".octave".  In bindings/octave/Makefile.am,
these files are listed (with the suffix) and there is a simple pattern rule
to "build" the %.octave files form their "sources" %.  These fles are then
installed in $(docdir).  That's simple and elegant, IMHO.

Also, there is now a unique place in configure.ac to define where the doc
files go: the AC_SUBSTituted DOCDIR variable.  This should be propagated to
other Makefile.am's in the source tree.

svn path=/trunk/; revision=4378