Joao has continuously complained that with the current PLplot in CVS it is not possible to use the library without doing make install.  He is right. One of the reasons for this is that the driver modules are looked for only under the absolute path DATA_DIR, which is actually a installation path.
I changed the code of plInitiDispatchTable in plcore.c to load the drivers
from the directory pointed by the env var PLPLOT_DRV_DIR. If this env var is
not defined, it uses the default DATA_DIR. This is a interim solution that
will allow test of PLplot (at least the access to driver modules) without a
full install.

There is also the problems of data files and fonts.  Some of these used to
be accessed in the past via the PLPLOT_LIB environment variable.  This has
to be sorted out and I would like to find a better global solution for this
problem of dual access (devel vs. install locations).

Another change in the code was the inclusion of some tests and plabort
calls.

svn path=/trunk/; revision=4401