Use AC_HELP_STRING whenever AC_ARG_{WITH,ENABLE} is involved.  This is the right way of doing things in order to have nicely formatted help strings. With this change, "configure --help" looks okay now.
There is still a glitch: some help strings show "$enable_<var>" instead of
the value of variable enable_<var>, like this:

  --enable-xwin           enable xwin device driver ($enable_xwin)

This has nothing to do with the current change and happened also before.  I
think that this is due to a recent change in Autoconf, where a delimiter of
a here-document string is quoted, preventing variable substitution.  My
guess is that there is nothing that we can do to fix that, but I will
investigate later.

svn path=/trunk/; revision=4406