This is a revival of the --with-rpath option to configure, which has been ripped out from sysloc.in when the AT branch has been created.  I need this to build Lintian clean packages in Debian.
Setting -rpath may be a handy thing for users who do not have root access,
but enforcing it by default is not a Good Thing to do.  In particular,
Debian Policy enforces the non-use of rpath for upgradability reasons.

This changes is transparent and harmless, and the PLplot build will happens
as before, since with_rpath is set to "yes" by default.  Please, do not undo
this change in the future!

svn path=/trunk/; revision=4418