Replaced     AM_LDFLAGS = $(libplplot) by     LDADD = $(libplplot)
Using LDADD instead of AM_LDFLAGS is the Right Way (TM) of doing things,
because the dependencies are generated correctly.  AM_LDFLAGS should only be
used to pass _flags_ to the link command, not library objects.

This reminds me that there are several other places in the source tree that
have the same problem.  They will be fixed progressively.

This commit also changed "../.."to "$(top_builddir)" in the definition of
the variable libplplot.  Again, this is the Right Way with Automake, since
it is more robust to eventual reorganizations of the source tree (okay, that
is not likely to happen with PLplot, but let us act in a principled way).

svn path=/trunk/; revision=4525