This is the brute force approach to get "make dist" to include (almost) all the files that are present in a freshly checked out CVS tree.  The notable exceptions are the directories cf/, new/, tmp/, and bindings/perl5. This was essentially done by putting files in the EXTRA_DIST variable in the Makefile.am files.
I really dislike this brute force method, but since we really need to make a
release soon, I see no way of doing otherwise.  I would prefer to take this
opportunity to remove some legacy cruft and to improve our installation in a
file-per-file basis.  For instance, in directory bindings/tcl, I added
blindly the files README.tclAPI and plitclgen to EXTRA_DIST, but it is clear
that they could be installed somewhere with "make install".  Same commentary
for the Python and Java stuff.  Since I am (almost) totally ignorant on
Python, Java and Tcl/Tk, I will really appreciate if competent developers
step in.

This cleanup + installation improvement discussion must take place in the
post-release times.

An aside note: the EXTRA_DIST variable (or the variables it uses) should not
be inside conditionals in Makefile.am, otherwise the distribution tarball
may be wrong.  This is not yet consistently done throughout the source tree.

svn path=/trunk/; revision=4545