Configuration changes so that make www-install is starting to work again, but there is more to do.
Also, fix prerequisite lists wherever I saw problems. (A more systematic
follow-up should be done.) This work involved correcting missing
dependencies as well as incorrect dependencies. (Often the *.in version of
the file was in the prerequsite list when clearly the file itself was being
used in the rule).

Also, drop make command when not needed.  I assume this was put in to
partially address the missing dependencies problems, but a better way to
do it is to fix the dependencies in the first place!

svn path=/trunk/; revision=4554