Made *.txt files in plplot_octave_txt be included in the distribution tarball, obtained through make dist.
The makefile rule to produce these files has been revamped, with right
dependencies to the files in doc/docbook/{bin,src}.  ALso, a time stamp file
(doc-stamp) is now generated (and also included in the tarball). The file
plplot_stub.m depends now on doc-stamp. The setup is such that for regular
users the rule should be never triggered.

The nice side-effect of this change is that regular users do not need the
Perl XML modules anymore in order to build the Octave buildings.  I have
still to add some logic to configure.ac to check for the Perl modules for
the developers who need to build the Octave documentation files.

svn path=/trunk/; revision=4558