Fixed a quite important bug: after cleaning the plotting function in ${prefix}/share/plplot_octave, the previous working directory was not restored.
Another less critical, but also important change: since this file is a
script, all the variables used in it are in the user's workspace.  Variables
names like "i", "t" or "path" were being used, which might conflict with
those of the users.  All the variables used in toggle_plplot_use.m are now
prefixed with "use_plplot_" and are cleared at the end of the script
(besides use_plplot_state, which must be kept between successive calls of
the script).

svn path=/trunk/; revision=4563