Implementation of Alan's idea for better control of Docbook documentation check/build.  The previous options --enable-docbook and --enable-docbook-build have been replaced by:
  --enable-builddoc       Enable configuration checks and build of DocBook
                          documentation. Default value is "no".

  --with-prebuiltdoc      Assume that the DocBook documentation is already
                          built and present in doc/docbook/src, ready to be
                          installed. This option is useful for package
                          maintainers. Default value is "no".

Specifying both options at the same time is logically inconsistent and will
result in an error message with configure been stopped.

The implementation was quite straightforward, because I just wrapped the two
AM_CONDITIONALs enable_docbook and docbook_build around the two configure
options above.

A nice improvement over the previous situation is that the Docbook checks
are not exercised when the docs are already present and needs only
installation (i.e. when using --with-prebuiltdoc).

svn path=/trunk/; revision=4679