This set of changes should fix some (if not all) of the remaining problems with the tarball generation related to built files generated with Perl.  The $PERL variable is now set the a warning message in configure.ac, such that if Perl in not found by configure, make will not die furiously.  I hope I traced down all the places where Perl scripts are invoked in the Makefile.am files and I prepended $(PERL) to the calls.  There may be some others remaining places still (excepted doc/docbook), though.
With the current changes, regular users building from the tarball do not
need Perl available in their systems in order to build and install PLplot.
The last problem that I noticed was with the generation of the
bindings/octave/plplot_octave_txt/*.txt files.  The doc-stamp target has
been brought outside the AM_CONDITIONAL enable_octave and the EXTRA_DIST
variable is not set conditionally anymore.

Also, fixed the macro CHECK_PORG in doc/docbook/docbook.m4, which is used to
detect the Perl program.  The variable PROG is now correctly set to the null
string if the program is not detected.

Unrelated to the above, a check for the libm is introduced before the check
of functions isinf, finite, and isnan.  This should fix the configure
problem in Alpha-OSF.

svn path=/trunk/; revision=4716