This is the workaround for the broken linker variables set in mkoctfile in the NetBSD system, as discussed in plplot-devel.
A new AC_SUSBTituted variable, called MKOCTFILE_VARS, has been created.  If
configure runs in a NetBSD* system and the C++ compiler is g++, then the
SH_LD and SH_LDFLAGS are set appropriately and used in the invokation of
mkoctfile in bindings/octave/Makefile.am.

A flashy warning comment has been put in configure.ac, and a AC_MSG_WARN is
issued when the variables are set.  It looks like this:

    configure: WARNING: Setting mkoctfile variables for broken NetBSD system

Of course, I cannot fully test this changes.  Please, Alan (or whoever),
test it.

svn path=/trunk/; revision=4724