In sysloc.in, changed the list of directories to search for Python headers (when setting the PYTHON_INC_DIR variable) to have just:
        $prefix/include/python$PYTHON_VERSION
        /usr/include/python$PYTHON_VERSION

instead of a long list of all possible Python versioned include dirs.

This change requires that the call to AM_PATH_PYTHON be made before
AC_INCLUDE(sysloc.in), in configure.in.

Also, an additional sanity check for non-emptiness of the PYTHON_VERSION
variable is added to sysloc.in, right at the beginning of the Python
configuration section.

svn path=/trunk/; revision=4831