AWI for Andrew Ross <andrewr@coriolis.greenend.org.uk>.  Andrew kindly donated these clones of the C/python/java/tcl examples under the LGPL.
I added the appropriate copyright line and (LGPL) copying conditions. I also
removed java commentary that had been inadvertently copied for x01.cc from
the java example.  I also temporarily disabled the HAVE_USLEEP code in
x01.cc and x17.cc which does not work on Linux (at least on my Debian stable
system).  Another issue is that x02, x04, x16, and x21 produce different
results than the corresponding C examples.  There is an obvious problem for
the second example, but the rest are subtle as in similar differences I
find between java and c.  The good news is that the rest of the
examples produce identical to C results which is a pretty good test that
the C++ interface has now entered the modern era.

Thanks Andrew!

svn path=/trunk/; revision=4851