AWI for Andrew Ross <andrewr@coriolis.greenend.org.uk> who kindly updated the C++ interface (under the LGPL) to bring it into the modern era.  At least 13 API additions were made.
N.B. the method of initializing plplot under the C++ interface has been
changed. So this is an API breakage, but it allows, for example, the user to
use command-line options for the first time for that interface.  Also, an
extensive set of C++ examples (clones of the corresponding C/python/java/tcl
examples) have been committed that use the updated C++ interface so it
should not be a difficult change for the user to make.

svn path=/trunk/; revision=4857