The robustness of the macro has been improved.  If GET_DLNAME is called before AC_LIBTOOL_PROG, the configure script will exit now with an appropriate error message.  This is not intended for general users, but only for developers who make changes to configure.ac.
Some comments about the last changes done by Alan:

In the GET_DLNAME macro, $LD was changed to $CC as the command when invoking
libtool --mode=link. Actually, the name of this command is irrelevant to the
actions taken by the libtool script.  However, in version 1.5 (or 1.5.0a) of
libtool, the name is used to guess a "tagged configuration", in case the
option --tag has not been specified.  Using $LD would then trigger an error.
I hope that this macro will be portable across different platforms. The
future will tell us.

svn path=/trunk/; revision=4886