Connected changes to drivers/Makefile.am, lib/nn/Makefile.am, and src/Makefile.am
Link xwin driver with libpthreads (or if disable dynamic drivers so that
xwin driver embedded in libplplot, link libplplot with libpthreads).

Link libnn with libqhull

Do not link libplplot with libqhull

Do not link libplplot with libpthreads (except for disabled dynamic drivers
case mentioned above and dealt with in the drivers directory).

svn path=/trunk/; revision=4892