Complete reorganization following what is done for ../../bindings/python/Makefile.am.
Note that install-exec-hook: is now used to compile all the installed java
files rather than compiling them in a local directory underneath
bindings/java before installation, and this reduces a lot of clutter that
was occurring before.

Another notable change is the java module that interfaces to PLplot is no
longer built like a full-blown library with a version number, precision tag,
etc. with installation into $prefix/lib with appropriate version number
symlinks.  Instead, it is built similarly to the python module and installed
in the location where the core java and compiled class files are installed.

svn path=/trunk/; revision=4923