Reintroduced deleted AC_MSG_RESULT in Nan awareness test
Also, the logic of the test is fixed now.  Strangely, the last commit made
AC_RUN_IFELSE have only two arguments, instead of three.  Besides that, the
test must return 1 in case of failure and not 0 as was the case.  I did
minimal tests, but evrything seems to work now.

svn path=/trunk/; revision=4942