If the current directory is in the build tree, change files search order so as to load them locally.
The modifications guarantee that files/drivers/scripts are loaded from the build tree if the current
directory is somewhere in the build tree.

The objective is that one should be able to change/test/evaluate PLplot without the necessity of
installing it; recently this was needed (and for some front ends it still is), requiring the user to have a
separate directory to test/change/evaluate PLplot, which means two steps of configure/make, one with
the test --prefix and the other with the final one, which is confusing and error prone. Also, for developpers,
the make/make install step turns the development cycle sloww and boring.

All the compiled front-ends, C, C++ and F77, was tested and work OK with all drivers.
The Tcl and Tk front-ends also work OK, but the tk loadable extensions, Pltk and Plplotter,
as well as the tcl extension Pltcl, still need adjustments in the pkgIndex.tcl script.

The Octave front-end also works OK.

Python and Java was not yet tested.

This is not yet finish, as some problems might yet appear; however, care has been taken to
guarantee that the old behavior still works OK.

tclAPI.c:
   PLbasicInit(): search the init script in the build tree
   pls_auto_path(): add bindings/tk to auto_path

plcore.c:
   add plInBuildTree() that check if the current directory in within the build source tree
   plGetDriverDir(): if in the build tree, load the drivers locally

plctrl.c:
    plFindCommand(): if in the build tree, search locally
    plLibOpenPdfstrm(): if in the build tree, search locally

svn path=/trunk/; revision=4943