Fixed Libs and Cflags entries
Added "/plplot" to the includedir variable, such that there is no need to
explicitly #include <plplot/...> in user programs (this is the case for the
C and C++ examples shiped with PLplot).

Also, removed -DPL_DOUBLE from Cflags (this was producing compilation
warnings, since PL_DOUBLE is already defined in plConfig.h) and removed
@FREETYPELIBCMD@ from Libs.  All the dependecy libraries of libplplot are
guessed by the default linker of Linux.  I am not sure this will work like
this in other Unices.  More work is needed here.

At any rate, compilation of the C examples work correctly now with:

    gcc `pkg-config plplot --cflags` x01c.c -o x01c `pkg-config plplot --libs`

svn path=/trunk/; revision=4947