Update and partial redesign of plplot-config, a simplified wrapper for plplot_libtool if you don't want to use the power of that script directly.
* fixed bugs where it failed for libtool-1.5 (a linker must be specified),
and also failed for --with-c++ (wrongly named library).

* introduce --with-f77 and --with-tcl options

* non-cumulative library modes.  The idea here is you ordinarily only tell
plplot_libtool about libplplot, and it figures out the correct link options
for the local platform based on that.  However, there are some additional
mutually exclusive options a user can specify:

--with-c++,
--with-f77, or
--with-tcl.

If you specify any of those additional switches you tell plplot_libtool only
about the appropriate library (libplplotcxx for --with-c++, libplplotf77 for
--with-f77, and libplplottcltk for --with-tcl), and it determines the
correct link options for your local platform.  The above switches should be
mutually exclusive, but my scripting skills aren't up to checking for that
(help, please!), but for now I settled for an increasing order of precedence
being in the above order (so anything lower on the list means anything above
is ignored).

svn path=/trunk/; revision=4950