pkg-config gives now complete list of dependency libraries
This is a first try to have "pkg-config plplot --libs" working correctly
in a cross-platform way.  For that, the complete list of dependency
libraries flags (-l*) must be included into plplot.pc and plplot-float.pc.

This information is already available in the src/.libs/libplplot*.lai files.
These files are parsed in the install-data-hook rule of pkgcfg/Makefile.am
in a similar way the plplot_libtool script does at run time.  However, this
is done at install time.

This is now the result I am having in Linux (must be also tested in other
platform):

    $ pkg-config plplot --libs
    -lplplotd -lfreetype -lz -lcsirocsa -lcsironn -lqhull -lm -ldl

svn path=/trunk/; revision=4951