Fixed problem with sed option -i
Option -i of the sed command is quite useful, because it allows files to be
edited in-place.  It is too bad that GNU sed older than version 4 do not
support this feature.  Since this should be the case for general sed's, the
use of option -i was replaced by an ugly (but working) code using a
temporary file.

svn path=/trunk/; revision=4956