Better logic for use of system's libltdl
The decision about when to use the libltdl library is improved.  Now, the
default is to always compiled the shipped libtool's libltdl as a convenience
library, unless the user explicitly give the option --with-ltdlsystem to
configure.  When this option is given, confgiure checks if the libltdl
library is actually installed.  If not, it falls back to the convenience
compilation case.

Alan, if you are hearing this, could you please comment?

svn path=/trunk/; revision=4959