Unversioned path for Octave oct dir
The OCTAVE_OCT_DIR variable is not versioned anymore (i.e. it does not
contain the Octave version 2.*.* in it).  This parallels what is done for
the OCT_M_DIR variable.  OCTAVE_OCT_DIR takes it value from the
octave_config_info structure, being either localoctfiledir (for Octave
version 2.1.*) or the first component of localoctfilepath (version 2.0.*).

The advantage of this change is that the Octave bindings are not tied to a
specific version of Octave, and will continue to be visible if the user
upgrades Octave.  Remember though that there is no guarantee of backward
compatibility of the *.oct API within the unstable series of Octave (2.1.*).

svn path=/trunk/; revision=4965