Populating the www/annouce directory
Here is everything that is needed for building XHTML, text, and PDF forms of
release annoucements from a DocBook-XML source.  File
announce-plplot-5.2.2.xml is just a boilerplate for now (inspired from
Alan's last release announcement).

CAVEAT: this is an experimental project and I have no concrete plans
regarding the schedule of the next release.

See the README file for more info.

svn path=/www/trunk/; revision=7518