Fixed debug and error messages.
In function plOpenFile, the message "Opened <filename>" is sent now through
pldebug.  This means that it will not spuriously appear on stderr anymore,
but will be issued when option -debug is specified.  Also, the message
"Can't open <filename>" is sent now to stderr instead of stdout.

In function plLibOpenPdfstr, two calls to pldebug are fixed now (called with
two argments instead of three).

svn path=/trunk/; revision=5023