Include most drivers/*.rc files in the distribution tarball
The goal of this change is to avoid the building of the drivers/*.rc files
by the users by including them into the distribution tarball.  This also
avoids the building of the get-drv-info utility when running "make" from the
tarball.

The price paid for this improvement on the user's side is that "make" is
needed before "make dist" (see scripts/make-cvs-tarball.sh). This happens
because get-drv-info needs libplplot, which needs libcsirocsa and libcsironn
in turn.  Also, libltdl needs to be built.

Since this change only affects the tarball construction and since the price
paid is too small compared with the gain obtained, I decided to implement
the new behavior. If anybody objects, please let us discuss in plplot-devel.

svn path=/trunk/; revision=5024