Disabled generation of reconfig script by configure
As per our discussion in plplot-devel in September 2003 (and since Maurice
agreed), the section in configure.ac responsible for the generation of the
reconfig script is removed.  The command "./config.status --recheck" has the
same effect (i.e. calling configure with the previously specified options).

In order to keep the functionality for the fingers too used to type
"./reconfig", a one-line shell script called reconfig has been added to the
repository.

Also, added reconfig to the EXTRA_DIST variable in the top-dir Makefile.am,
such that the script gets included into the distribution tarball.

svn path=/trunk/; revision=5025