Disabled inclusion of drivers/*.rc in the distribution tarball
Revert the previous changes regarding includion of the drivers/*.rc files
into the tarball, because a .rc file depend on the .la file, which gets
built with "make" by the user.  The result is that the *.rc and get-drv-info
are built anyway.

I will think later on a better way to include the *.rc files in the tarball.

svn path=/trunk/; revision=5027