Style changes: comments, do loops, indentation, and plcol ==> plcol0.
Change 15th example to (a) make it work for the first time (there were
incorrect arguments in some cases, and the 15th example uses more than 3
colours in cmap0 so that is why plscmap0n(3) is commented out in the C
example case and is now dropped altogether in this fortran example case) and
(b) obtain exact agreement with file results from the corresponding C
example. This agreement confirms that the fortran interface plshade0
function is implemented correctly.

svn path=/trunk/; revision=5043