Sane treatment of DATA_DIR and DRV_DIR
The variables/#defines DATA_DIR and DRV_DIR are treated now in a more
coherent and Autotools-friendly way.  The main changes regard:

    - Use of $libdir and $datadir as prefixes of DRV_DIR and DATA_DIR,
      respectively, in configure.ac.  This allows users to easily override
      the installation locations when running configure (options --libdir
      and --datadir).  The environment variables DATA_DIR and DRV_DIR are
      still honored.

    - DRV_DIR is an absolute path now, not relative to DATA_DIR as before.
      Besides it being cleaner and more intuitive, the new design allows
      better compliance with the FHS.  Indeed, architecture-independent
      files (like the font amd map files, as weel as examples and the Tcl
      stuff) go by default in $prefix/share, while driver module files
      (which are architecture-dependent) go in $prefix/lib.

As regards final results, the only difference with the previous situation is
that the *.fnt and *.map files are installed now in plplot<version>/ instead
of plplot<version>/data.  I did not try to keep the previous situation
because it would imply in the definition of a new variable/#define.
However, if the other developers have really strong arguments in favor of
it, I will implement that.

svn path=/trunk/; revision=5061