Fix for the "centuries old" very irritating bogus warning messages about itclDecls.h during configuration (this should improve your day :).
The problem was, itclDecls.h apparently is not meant to be included except
from itcl.h, or after itcl.h has been included first.  So, this dependency
needed to be accounted for, and fortunately there is an autoconf construct
to do it.  As is often the case with these things, very simple once you know
what to do.

svn path=/trunk/; revision=5111