Restructured the pkg-config support.
Everything that was scattered across pkgcfg/gen-pc-files.sh and
pkgcfg/Makefile.am (shell code) as well as pkgcfg/plplot-master-pc.in (.pc
file template) is now regrouped in the single Perl script
gen-pc-files.pl.in, configurable through configure.

Since Perl is needed now, the code in configure.ac was changed, such that
pkg-config support is disabled if Perl cannot found in the system. Also the
pkg-config section in configure.ac was moved after the call to
CHECK_PROG(perl).

The Tcl/Tk specific cflags for include directories are now integrated in
plplot*-tcl.pc.  The file examples/tk/Makefile.examples.in was changed
accordingly.

svn path=/trunk/; revision=5124