Fortran API change as discussed on the list.
All plshade[0-2] and plshade fortran calls now have a last parameter which
is the first dimension of the 2D arrays.  This extra parameter is similar to
what is currently done for the fortran versions of plmesh, plmeshc, plot3d,
plot3dc, plshades[0-2], and plshades.  It allows large static 2D arrays to
be dimensioned in fortran apps which call these PLplot functions (plcon[0-2]
and plcont have similar possibilities for large static arrays done a
different way).  This in turn allows smaller areas of the fortran array to
be populated as needed (interactively from input, say) without having to
recompile for every change in size of the defined part of the array.

These recent Fortran API changes have been tested for fortran examples 15
and 16a (now both changes to use static arrays dimensioned larger than the
defined area), and identical results to the previous are obtained.
Furthermore, memory management has been changed substantially for fortran
plshade[0-2] and plshade, and valgrind applied to fortran examples 15 and
16a show no problems with these memory management changes.

svn path=/trunk/; revision=5127