Rafael's script-based method of getting and using libtool information to solve mkoctfile linking issues and a more careful approach to overcoming the run library path issues of mkoctfile has proved successful not only on Linux but also cross-platform. Therefore, abandon the pure libtool-based approach to build the octave interface.
Note, we use the pure libtool-based approach to build every other interface,
and building the octave interface worked fine with the pure libtool approach
in the Linux case, but there was one cross-platform showstopper (couldn't
get the -shrext libtool option to work in the documented manner to set the
shared object suffix to .oct as demanded by octave) that I hadn't solved yet
when Rafael beat me to a solution using a special script-based method.

Start here, if future developers want to use the pure libtool approach for
building the octave interface.

svn path=/trunk/; revision=5156