This jumbo cvs commit is the result of a herculean task to update the copyright notices of most of the source files of PLplot.
I hope that I did not introduce any serious breakage.  This seems plausible,
because I did changes only at the beginning of the files and only in
commented lines.  However, I used a semi-automatic procedure for doing that
and some bugs may have slipped into the files.  The sources still build okay
(at least the Debian packages) and I can install the libraries, and compile
and run some C examples.  That is not enough, so PLEASE TEST!

This is what I did in more detail:

1) From the CVS ChangeLog, I considered all the files that changed since the
   last release (5.2.1).  I then added the line:

       Copyright (C) 2004  Joe C. V. S. Committer

   where the name of the developper who did the CVS commit is included.

2) Notice that the year in the line above should read 2004, even if the
   changes were done last year, since Copyright years refer to when the
   software is released.

3) For the same reason, I did not bother adding missing Copyright entries
   for previous authors of each file.  One copyright holder is enough for
   our purposes.

4) I also reformatted and reworded the License Conditions text.  This
   included changing things like "This file is free software" to "PLplot is
   free sofware" and so on.

5) Finally, some files were lacking License notices, especially the src/*.c.
   I added them and put in the Copyright holder line the name of the
   developer who did the last cvs commit for the file.

If you wish to know where your name appears, you can run from the
$(top_builddir):

    grep -ri "copyright.*2004".

svn path=/trunk/; revision=5160