AWI for Andrew Ross memory management cleanup.
Add #ifndef ENABLE_DYNDRIVERS ... #endif about the lines in
plD_dispatch_init_XXXXX which assign pdt->pl_MenuStr and pdt->pl_DevName. If
this is a dynamic driver then we have already assigned these strings in
plcore.c using the database. We absolutely don't want to use a static string
in the dynamic driver to initialise it with since that string will disappear
once the driver is unloaded and we are left with no valid driver name or
menu item. Oh and overwriting the string will also leak the previously
allocated memory for the names.

svn path=/trunk/; revision=5181