[ Forcing cvs commit to get the correct log message, screwed up in the   previous commit: ]
When PLplot is configured with disable_shared, dyndrivers must also be
disabled. However, this test was done too late in configure.ac, after the
libltdl checks were done conditionally on the value of enable_dyndrivers.
This was causing needless checks and configuration of libltdl.

Since this check is done now before AC_PROG_LIBTOOL, where enable_shared is
usually initialized, an explicit initialization of enable_shared (=yes)  was
needed at the beginning of configure.ac.

svn path=/branches/unlabeled-1.150.2/; revision=5191