Use a PHP function instead of copy-and-paste code repeated in all demo*/index.php files.  This will simplify maintenance in teh future, since the code is now concentrated in a single place.
The results should be identical to what we had before.  The index.php files
look now like this (e.g. for demo 16):

    <?php
    include ("../example_page.php");
    demo_page (16, 5, $lbind);
    ?>

The first two arguments of the demo_page function (defined in the new PHP
file examples/example_page.php) are the demo number and the number of plot
images (x*.*.png files).  The third argument must always be "$lbind".  That
is ugly and I will find a better solution later.  Please notice that demo20
produces a wrong result.  I will fix this eventually.

Also, the demo files will be at the htdocs root directory instead of in the
examples directory.

svn path=/www/trunk/; revision=7552