Test of automatic inclusion of macros defined in files cf/*.m4 by aclocal. The file cf/gtk.m4 is added to the project.  This will be the canonical way to add third-party m4 macros file here.
Also, get the PL_INCLUDE macro working.  It includes file from the cf/ dir.
For now, we have sysloc.in and docbook.in, but these will soon disappear.

The only thing that I do not fully understand yet is why aclocal-1.8 does
not follow the inclusions of other files in configure.ac.  Hence the
AC_REQUIRE([AM_PATH_GTK]).  It seems that mentions to external macros must
be mentioned in configure.ac and not in included files.  This will need more
study in the future, but AC_REQUIRE works for now.

svn path=/branches/CFDIR/; revision=5368